+++ src/DataStructures/Tensor/Tensor.hpp
@@ -268,7 +268,7 @@ class Tensor<X, Symm, IndexList<Indices...>> {
   /// Retrieve a TensorExpression object with the index structure passed in
   template <typename... N,
             Requires<std::conjunction_v<tt::is_tensor_index<N>...> and
-                     tmpl::is_set<tmpl::int32_t<N::value>...>::value> = nullptr>
+                     tmpl::is_set<tmpl::size_t<N::value>...>::value> = nullptr>
   SPECTRE_ALWAYS_INLINE constexpr TE<tmpl::list<N...>> operator()(
       N... /*meta*/) const noexcept {
     return TE<tmpl::list<N...>>(*this);
@@ -277,27 +277,27 @@ class Tensor<X, Symm, IndexList<Indices...>> {
   template <
       typename... N,
       Requires<std::conjunction_v<tt::is_tensor_index<N>...> and
-               not tmpl::is_set<tmpl::int32_t<N::value>...>::value> = nullptr>
+               not tmpl::is_set<tmpl::size_t<N::value>...>::value> = nullptr>
   SPECTRE_ALWAYS_INLINE constexpr auto operator()(N... /*meta*/) const noexcept
       -> decltype(
           TensorExpressions::fully_contracted<
               TE,
               replace_indices<
                   tmpl::list<N...>,
-                  repeated<tmpl::list<tmpl::int32_t<N::value>... /*N...*/>>>,
-              tmpl::int32_t<0>,
+                  repeated<tmpl::list<tmpl::size_t<N::value>... /*N...*/>>>,
+              tmpl::size_t<0>,
               tmpl::size<repeated<tmpl::list<
-                  tmpl::int32_t<N::value>... /*N...*/>>>>::apply(*this)) {
+                  tmpl::size_t<N::value>... /*N...*/>>>>::apply(*this)) {
     using args_list = tmpl::list<N...>;
-    using tensorindex_values = tmpl::list<tmpl::int32_t<N::value>...>;
+    using tensorindex_values = tmpl::list<tmpl::size_t<N::value>...>;
     // using repeated_args_list = repeated<args_list>;
     using repeated_tensorindex_values = repeated<tensorindex_values>;
     /*return TensorExpressions::fully_contracted<
-        TE, replace_indices<args_list, repeated_args_list>, tmpl::int32_t<0>,
+        TE, replace_indices<args_list, repeated_args_list>, tmpl::size_t<0>,
         tmpl::size<repeated_args_list>>::apply(*this);*/
     return TensorExpressions::fully_contracted<
         TE, replace_indices<args_list, repeated_tensorindex_values>,
-        tmpl::int32_t<0>,
+        tmpl::size_t<0>,
         tmpl::size<repeated_tensorindex_values>>::apply(*this);
   }
   // @}
